Ensure that a User model is saved if possible
A form that submits a `user_pk` should cause a User model to get saved
and associated with the `Donor`.  For some reason, this doesn't get
saved automatically inside the `DonorForm`, so I'm doing it manually
here.  There's probably something that can be tweaked to make it work
without the extra logic.